Extract dashboard view composition
